requirements.txt
openai
pandas
tqdm
python-dotenv
cpi
matplotlib
seaborn
requests 